서치 통합, navbar 반응형으로 만들기
