+++ package.json
@@ -3,15 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
     "@date-io/date-fns": "^2.11.0",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
     "@formatjs/intl": "2.10.4",
     "@mui/base": "5.0.0-beta.40",
     "@mui/icons-material": "5.15.20",
-    "@mui/material": "^5.15.20",
     "@mui/lab": "^5.0.0-alpha.170",
     "@mui/material": "5.15.20",
     "@mui/styled-engine-sc": "6.0.0-alpha.18",
@@ -27,6 +24,7 @@
     "cloudinary-react": "^1.8.1",
     "date-fns": "^2.30.0",
     "dayjs": "^1.11.11",
+    "material-ui-popup-state": "5.1.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-kakao-login": "^2.1.1",
